+++ xcm-simulator/src/laosish/configs/mod.rs
@@ -1,0 +1,1 @@
+mod balances;
